generank/api/serializers/news_feed.py
from django.core.exceptions import ObjectDoesNotExist
from rest_framework.reverse import reverse
from rest_framework import serializers
from rest_framework import status
from rest_framework.response import Response

from ..models import Item


class ItemSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Item
        fields = ('link', 'title', 'description', 'image', 'created_on')
        extra_kwargs = {'url': {'view_name': 'api:item-detail'}}
