requirements.txt
amqp==2.1.4
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.5.0.2
celery==4.0.2
cycler==0.10.0
defusedxml==0.4.1
Django==1.9.5
django-allauth==0.25.2
django-braces==1.9.0
django-celery-beat==1.0.1
django-extensions==1.6.1
django-filter==0.13.0
django-oauth-toolkit==0.10.0
django-push-notifications==1.4.1
django-sslserver==0.19
djangorestframework==3.3.3
drf-nested-routers==0.11.1
ecdsa==0.13
et-xmlfile==1.0.1
Fabric3==1.10.2.post3
gunicorn==19.6.0
iron-cache==0.3.2
iron-celery==0.4.5
iron-core==1.2.0
iron-mq==0.8
jdcal==1.3
kombu==4.0.2
Markdown==2.6.6
matplotlib==2.0.0
mock==2.0.0
nose-cov==1.6
numpy==1.12.0
oauthlib==1.0.3
openpyxl==2.3.5
paramiko==1.16.0
pbr==1.10.0
praw==4.5.1
prawcore==0.10.1
psycopg2==2.6.1
pycrypto==2.6.1
pyparsing==2.1.10
python-dateutil==2.5.3
python-gnupg==0.3.8
python3-openid==3.0.10
python-opengraph==0.0.2
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
rest-condition==1.0.2
six==1.10.0
sqlparse==0.1.19
update-checker==0.16
vine==1.1.3
wheel==0.24.0
